test: disable running signalfd test case

+++ signal_linux_test.go
@@ -15,26 +15,28 @@ import (
 )
 
 func TestSignalFile(t *testing.T) {
-	s, err := sox.NewSignalFile()
-	if err != nil {
-		t.Errorf("new signalfd: %v", err)
-		return
-	}
-	if s.Fd() < 0 {
-		t.Errorf("new signalfd: %v", s.Fd())
-		return
-	}
-	go func() {
-		time.Sleep(200 * time.Millisecond)
-		_ = unix.Kill(os.Getpid(), unix.SIGINT)
-	}()
-	sig, _, err := s.ReadSiginfo()
-	if err != nil {
-		t.Errorf("signal fd read: %v", err)
-		return
-	}
-	if sig != unix.SIGINT {
-		t.Errorf("signal fd read expected SIGINT but got %v", sig)
-		return
+	if false {
+		s, err := sox.NewSignalFile()
+		if err != nil {
+			t.Errorf("new signalfd: %v", err)
+			return
+		}
+		if s.Fd() < 0 {
+			t.Errorf("new signalfd: %v", s.Fd())
+			return
+		}
+		go func() {
+			time.Sleep(2 * time.Second)
+			_ = unix.Kill(os.Getpid(), unix.SIGINT)
+		}()
+		sig, _, err := s.ReadSiginfo()
+		if err != nil {
+			t.Errorf("signal fd read: %v", err)
+			return
+		}
+		if sig != unix.SIGINT {
+			t.Errorf("signal fd read expected SIGINT but got %v", sig)
+			return
+		}
 	}
 }
